requirements.txt
termcolor==1.1.0
jsonschema==2.6.0
setuptools==40.0.0
tabulate==0.8.2
uperations_kernel==0.0.7
PyContracts==1.8.3
python-dotenv==0.9.1
kernel==0.0.0
PyYAML==3.13
